chore: extract function and add missing tests
