+++ packages/renderer/src/lib/statusbar/StatusBarItem.svelte
@@ -1,26 +1,8 @@
 <script lang="ts">
 import type { StatusBarEntry } from '../../../../main/src/plugin/statusbar/statusbar-registry';
+import { iconClass } from './StatusBarItem';
 
 export let entry;
-
-function iconClass(entry: StatusBarEntry): string | undefined {
-  let iconClass = undefined;
-  if (entry.enabled && entry.activeIconClass !== undefined && entry.activeIconClass.trim().length !== 0) {
-    iconClass = entry.activeIconClass;
-  } else if (!entry.enabled && entry.inactiveIconClass !== undefined && entry.inactiveIconClass.trim().length !== 0) {
-    iconClass = entry.inactiveIconClass;
-  }
-  // handle ${} in icon class
-  // and interpret the value and replace with the class-name
-  if (iconClass !== undefined) {
-    const match = iconClass.match(/\$\{(.*)\}/);
-    if (match !== null && match.length === 2) {
-      const className = match[1];
-      iconClass = iconClass.replace(match[0], `podman-desktop-icon-${className}`);
-    }
-  }
-  return iconClass;
-}
 
 function tooltipText(entry: StatusBarEntry): string {
   return entry.tooltip !== undefined ? entry.tooltip : '';
